+++ libs/sdk/testing/private/src/a11y/a11y-analyzer-config.ts
@@ -1,7 +1,7 @@
 /**
  * @internal
  */
-export interface SkyA11yAnalyzerConfig {
+export interface _SkyA11yAnalyzerConfig {
   rules: Record<
     string,
     {