+++ libs/sdk/testing/src/index.ts
@@ -1,8 +1,7 @@
+// Export the a11y analyzer to avoid a breaking change. Remove in a future major version.
 // eslint-disable-next-line @nx/enforce-module-boundaries
-export {
-  SkyA11yAnalyzer,
-  type SkyA11yAnalyzerConfig,
-} from '@skyux-sdk/testing/private';
+export { _SkyA11yAnalyzer as SkyA11yAnalyzer } from '@skyux-sdk/testing/private';
+export type { SkyA11yAnalyzerConfig } from './lib/matchers/a11y-analyzer-config';
 export {
   SkyAsyncMatchers,
   SkyMatchers,